+++ Sources/NablaCore/Utils/Extensions/UIKit/UINavigationController.swift
@@ -1,15 +1,14 @@
 import Foundation
 import UIKit
 
-public extension NablaExtension where Base == UINavigationController {
+public extension NablaExtension where Base: UINavigationController {
     @discardableResult
     func withOpaqueNavigationBarBackground() -> Base {
-        let appearance = UINavigationBarAppearance()
-        appearance.configureWithOpaqueBackground()
-        base.navigationBar.standardAppearance = appearance
-        base.navigationBar.scrollEdgeAppearance = appearance
-        base.navigationBar.compactAppearance = appearance
-
+        let scrollEdgeAppearance = base.navigationBar.scrollEdgeAppearance
+            ?? UINavigationBar.appearance().scrollEdgeAppearance?.copy()
+            ?? UINavigationBarAppearance()
+        scrollEdgeAppearance.configureWithOpaqueBackground()
+        base.navigationBar.scrollEdgeAppearance = scrollEdgeAppearance
         return base
     }
 }