+++ Sources/NablaScheduling/Views/Scenes/CategoryPicker/CategoryPickerViewTheme.swift
@@ -12,6 +12,8 @@ public extension NablaTheme {
         public static var emptyViewFont: UIFont = Fonts.caption
         
         public enum CellTheme {
+            /// Corner radius used for the background of a category cell.
+            public static var cornerRadius = CGFloat(12)
             /// Color of the background of the cell of a category.
             public static var backgroundColor: UIColor = Colors.Fill.card
             /// Color of the text displayed with the name of a category.